Move `:stdlib-kotlin-extensions` to internal package
And deprecate `capitalized` extension, accidentally used by the nowInAndroid
build.
