+++ platforms/core-configuration/configuration-cache/src/main/kotlin/org/gradle/configurationcache/EncryptionService.kt
@@ -19,7 +19,7 @@ package org.gradle.configurationcache
 import org.gradle.cache.CacheBuilder
 import org.gradle.cache.PersistentCache
 import org.gradle.cache.scopes.GlobalScopedCacheBuilderFactory
-import org.gradle.configurationcache.extensions.useToRun
+import org.gradle.internal.extensions.stdlib.useToRun
 import org.gradle.configurationcache.initialization.ConfigurationCacheStartParameter
 import org.gradle.internal.hash.HashCode
 import org.gradle.internal.hash.Hashing
