+++ platforms/core-configuration/configuration-cache/src/main/kotlin/org/gradle/configurationcache/initialization/ConfigurationCacheStartParameter.kt
@@ -19,7 +19,7 @@ package org.gradle.configurationcache.initialization
 import org.gradle.StartParameter
 import org.gradle.api.internal.StartParameterInternal
 import org.gradle.api.logging.LogLevel
-import org.gradle.configurationcache.extensions.unsafeLazy
+import org.gradle.internal.extensions.stdlib.unsafeLazy
 import org.gradle.configurationcache.serialization.Workarounds
 import org.gradle.initialization.StartParameterBuildOptions.ConfigurationCacheProblemsOption
 import org.gradle.initialization.layout.BuildLayout
