+++ platforms/core-configuration/configuration-cache/src/main/kotlin/org/gradle/configurationcache/problems/ConfigurationCacheProblemsSummary.kt
@@ -22,8 +22,8 @@ import com.google.common.collect.Ordering
 import it.unimi.dsi.fastutil.objects.ObjectOpenHashSet
 import org.gradle.api.Task
 import org.gradle.api.internal.DocumentationRegistry
-import org.gradle.configurationcache.extensions.capitalized
 import org.gradle.internal.configuration.problems.PropertyProblem
+import org.gradle.internal.extensions.stdlib.capitalized
 import org.gradle.internal.logging.ConsoleRenderer
 import java.io.File
 import java.util.Comparator.comparing
