+++ platforms/core-configuration/configuration-cache/src/test/kotlin/org/gradle/configurationcache/problems/JsonModelWriterTest.kt
@@ -17,7 +17,7 @@
 package org.gradle.configurationcache.problems
 
 import groovy.json.JsonSlurper
-import org.gradle.configurationcache.extensions.uncheckedCast
+import org.gradle.internal.extensions.stdlib.uncheckedCast
 import org.gradle.internal.configuration.problems.DecoratedPropertyProblem
 import org.gradle.internal.configuration.problems.PropertyTrace
 import org.gradle.internal.configuration.problems.StructuredMessage
